pyproject.toml
[project]
name = "tela"
version = "0.2.0"
description = "Simple tools for teaching programming with graphics"
readme = "readme.md"
license = { text = "GNU GPL 3.0" }
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "pygame-ce>=2.5.2,<3",
]
authors = [
    { name = "Nilo Menezes", email = "nilo@nilo.pro.br" }
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
